refactor: EventHandler이름 수정
